requirements.txt
flask==3.0.3
flask-cors==5.0.0
langchain-core
langchain-openai
langgraph
python-dotenv
requests
pydantic
langgraph-checkpoint-postgres
langchain-community
psycopg-pool==3.2.3
pygraphviz
gunicorn==21.2.0